+++ vitest.workspace.mjs
@@ -2,4 +2,5 @@ import { defineWorkspace } from "vitest/config"
 
 export default defineWorkspace([
   "packages/*/vitest.config.mjs",
+  "tools/csp/vitest.config.mjs",
 ])