+++ packages/libjpeg-turbo-12bit/CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.4](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.3...@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.4) (2026-09-01)
+
+**Note:** Version bump only for package @cornerstonejs/codec-libjpeg-turbo-12bit
+
+
+
+
+## [0.4.3](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.2...@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.3) (2026-08-13)
+
+**Note:** Version bump only for package @cornerstonejs/codec-libjpeg-turbo-12bit
+
+
+
+
+
 ## [0.4.2](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.1...@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.2) (2026-07-09)
 
 **Note:** Version bump only for package @cornerstonejs/codec-libjpeg-turbo-12bit