+++ packages/little-endian/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.8](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-little-endian@0.0.7...@cornerstonejs/codec-little-endian@0.0.8) (2026-09-01)
+
+**Note:** Version bump only for package @cornerstonejs/codec-little-endian
+
+
+
+
 ## [0.0.7](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-little-endian@0.0.6...@cornerstonejs/codec-little-endian@0.0.7) (2026-07-09)
 
 **Note:** Version bump only for package @cornerstonejs/codec-little-endian